requirements.txt
playwright==1.58.*
python-dotenv
fastapi
uvicorn[standard]
hvac
